Workaround for potential rounding problems
For the ext led light (next to the colour sensor) there is a problem with led_off as the rounding sets the
fade out and fade in times to be non-zero. This results in the driver executing the LED_DRV_ENABLED_SINGLE_SHOT
branch of the case statement and trying to call app_timer_start with a timeout value of 0. This is turn causes
an error to be returned.

Workaround puts the LED into breath mode (not a good solution but at least it compiles) while I can investigate
the differences in the round function.